+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/_global/validation/annotation/ValidNickname.java
@@ -1,6 +1,6 @@
-package git_kkalnane.backend.starbucks.member.validation.annotation;
+package git_kkalnane.backend.starbucks._global.validation.annotation;
 
-import git_kkalnane.backend.starbucks.member.validation.validator.ValidNicknameValidator;
+import git_kkalnane.backend.starbucks._global.validation.validator.ValidNicknameValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 import java.lang.annotation.Documented;
