+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/_global/validation/validator/ValidPasswordValidator.java
@@ -1,6 +1,6 @@
-package git_kkalnane.backend.starbucks.member.validation.validator;
+package git_kkalnane.backend.starbucks._global.validation.validator;
 
-import git_kkalnane.backend.starbucks.member.validation.annotation.ValidPassword;
+import git_kkalnane.backend.starbucks._global.validation.annotation.ValidPassword;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
